Rename `neon::types::internal` to `neon::types::private`.
